+++ src/components/Textbox.mjs
@@ -34,7 +34,7 @@ export const Textbox = React.forwardRef(
             border-radius: 5px;
             padding: ${PADDING_EM}em;
             font: inherit;
-            font-size: 1.15rem;
+            font-size: 115%;
             font-style: italic;
             line-height: var(--daui-line-height);
             color: inherit;