+++ src/styles/stylesHtml.mjs
@@ -4,10 +4,10 @@ import css from 'styled-jsx/css'
 
 export const stylesHtml = css.resolve`
   & :global(h1) {
-    font-size: calc(2rem + 3vw);
+    font-size: calc(200% + 3vw);
   }
 
   & :global(h2) {
-    font-size: calc(1.5rem + 1.5vw);
+    font-size: calc(150% + 1.5vw);
   }
 `